V1.3

+++ Espera.h
@@ -3,7 +3,7 @@
 //  Sistema de espera sin usar timer
 //
  // (c)  José Angel Moneo (2016)
-// V1.2
+// V1.3
 
 
 #ifndef _Espera_class_H
@@ -22,14 +22,14 @@ class TEspera
 
  private:
        unsigned long Tsec;
-       unsigned int Tespera; //tiempo almacenado
+       unsigned long Tespera; //tiempo almacenado
        boolean fdone=true; //marca tiempo finalizado, para no tener que comprobar el tiempo despues de que se cumpla
  
 public:
    
    TEspera(){}; //constructor para servo. indicamos el las salidas del servo
    boolean done(bool op=false);
-   void set(unsigned int espera);
+   void set(unsigned long espera);
    void reset(); //Reinicia el tiempo preficado anteriormente
 
 	};
